+++ sys_zerocopy_linux.go
@@ -30,7 +30,7 @@ func setZeroCopy(fd int) error {
 
 func setBlockZeroCopySend(fd int, sec, usec int64) error {
 	return syscall.SetsockoptTimeval(fd, syscall.SOL_SOCKET, SO_ZEROBLOCKTIMEO, &syscall.Timeval{
-		Sec:  sec,
-		Usec: usec,
+		Sec:  int32(sec),
+		Usec: int32(usec),
 	})
 }
